Differenciate end of input failures from other failures.
For partial input parsing, we need to know when failure is caused by missing
data so that we can retry parsing later when we have more data.

This commit adds a need_data boolean flag alongs the failed one. It adds a stop
function to set that flag and propagate the failure. It replaces checks for
failure with a check on both failed and need_data flags. It also adds a eos
function to check for end of input. It also replaces the fail calls due to
end of input by stop calls.