Split json value parse function.
For partial support, when continuing parsing, we need to restart in the same
state as we left. For that, we need to skip some actions if they were already
done. Left as is, the code would be littered with ifs so refactor the parse_json
function into multiple function, one per type of objects.